Add Restore job and update Cleanup Jobs method

+++ src/cyborgbackup/main/management/commands/cleanup_jobs.py
@@ -1,3 +1,16 @@
+import os
+import re
+import stat
+import tempfile
+from io import StringIO
+from collections import OrderedDict
+from django.conf import settings
+from elasticsearch import Elasticsearch
+from distutils.version import LooseVersion as Version
+from cyborgbackup.main.expect import run
+from cyborgbackup.main.models.settings import Setting
+from cyborgbackup.main.utils.common import get_ssh_version
+from cyborgbackup.main.utils.encryption import decrypt_field
 
 # Django
 from django.core.management.base import BaseCommand
@@ -6,13 +19,24 @@ from django.db import transaction
 # CyBorgBackup
 from cyborgbackup.main.models import Job, Repository
 
+es_conf = settings.ELASTICSEARCH_DSL['default']['hosts'].split(':')
+es = Elasticsearch([{'host': es_conf[0], 'port': int(es_conf[1])}])
+
+OPENSSH_KEY_ERROR = u'''\
+It looks like you're trying to use a private key in OpenSSH format, which \
+isn't supported by the installed version of OpenSSH on this instance. \
+Try upgrading OpenSSH or providing your private key in an different format. \
+'''
+
 
 class Command(BaseCommand):
     '''
     Management command to cleanup old jobs.
     '''
 
-    help = 'Remove old jobs  from the database.'
+    help = 'Remove old jobs from the database.'
+
+    cleanup_paths = []
 
     def add_arguments(self, parser):
         parser.add_argument('--dry-run', dest='dry_run', action='store_true',
@@ -21,6 +45,142 @@ class Command(BaseCommand):
         parser.add_argument('--jobs', dest='only_jobs', action='store_true',
                             default=True,
                             help='Remove jobs')
+
+    def get_password_prompts(self, **kwargs):
+        d = OrderedDict()
+        for k, v in kwargs['passwords'].items():
+            d[re.compile(r'Enter passphrase for .*'+k+r':\s*?$', re.M)] = k
+            d[re.compile(r'Enter passphrase for .*'+k, re.M)] = k
+        d[re.compile(r'Bad passphrase, try again for .*:\s*?$', re.M)] = ''
+        return d
+
+    def get_ssh_key_path(self, instance, **kwargs):
+        '''
+        If using an SSH key, return the path for use by ssh-agent.
+        '''
+        private_data_files = kwargs.get('private_data_files', {})
+        if 'ssh' in private_data_files.get('credentials', {}):
+            return private_data_files['credentials']['ssh']
+
+        return ''
+
+    def build_passwords(self, job, **kwargs):
+        '''
+        Build a dictionary of passwords for SSH private key, SSH user, sudo/su.
+        '''
+        passwords = {}
+        for setting in Setting.objects.filter(key__contains='ssh_key'):
+            set = Setting.objects.get(key=setting.key.replace('ssh_key', 'ssh_password'))
+            passwords['credential_{}'.format(setting.key)] = decrypt_field(set, 'value')
+        return passwords
+
+    def build_private_data(self, instance, **kwargs):
+        '''
+        Return SSH private key data (only if stored in DB as ssh_key_data).
+        Return structure is a dict of the form:
+        '''
+        private_data = {'credentials': {}}
+        for sets in Setting.objects.filter(key__contains='ssh_key'):
+            # If we were sent SSH credentials, decrypt them and send them
+            # back (they will be written to a temporary file).
+            private_data['credentials'][sets] = decrypt_field(sets, 'value') or ''
+
+        return private_data
+
+    def build_private_data_dir(self, instance, **kwargs):
+        '''
+        Create a temporary directory for job-related files.
+        '''
+        path = tempfile.mkdtemp(prefix='cyborgbackup_%s_' % instance.pk, dir='/tmp/')
+        os.chmod(path, stat.S_IRUSR | stat.S_IWUSR | stat.S_IXUSR)
+        self.cleanup_paths.append(path)
+        return path
+
+    def build_private_data_files(self, instance, **kwargs):
+        '''
+        Creates temporary files containing the private data.
+        Returns a dictionary i.e.,
+
+        {
+            'credentials': {
+                <cyborgbackup.main.models.Credential>: '/path/to/decrypted/data',
+                <cyborgbackup.main.models.Credential>: '/path/to/decrypted/data',
+                <cyborgbackup.main.models.Credential>: '/path/to/decrypted/data',
+            }
+        }
+        '''
+        private_data = self.build_private_data(instance, **kwargs)
+        private_data_files = {'credentials': {}}
+        if private_data is not None:
+            ssh_ver = get_ssh_version()
+            ssh_too_old = True if ssh_ver == "unknown" else Version(ssh_ver) < Version("6.0")
+            openssh_keys_supported = ssh_ver != "unknown" and Version(ssh_ver) >= Version("6.5")
+            for sets, data in private_data.get('credentials', {}).items():
+                # Bail out now if a private key was provided in OpenSSH format
+                # and we're running an earlier version (<6.5).
+                if 'OPENSSH PRIVATE KEY' in data and not openssh_keys_supported:
+                    raise RuntimeError(OPENSSH_KEY_ERROR)
+            listpaths = []
+            for sets, data in private_data.get('credentials', {}).items():
+                # OpenSSH formatted keys must have a trailing newline to be
+                # accepted by ssh-add.
+                if 'OPENSSH PRIVATE KEY' in data and not data.endswith('\n'):
+                    data += '\n'
+                # For credentials used with ssh-add, write to a named pipe which
+                # will be read then closed, instead of leaving the SSH key on disk.
+                if sets and not ssh_too_old:
+                    name = 'credential_{}'.format(sets.key)
+                    path = os.path.join(kwargs['private_data_dir'], name)
+                    run.open_fifo_write(path, data)
+                    listpaths.append(path)
+            if len(listpaths) > 1:
+                private_data_files['credentials']['ssh'] = listpaths
+            elif len(listpaths) == 1:
+                private_data_files['credentials']['ssh'] = listpaths[0]
+
+        return private_data_files
+
+    def launch_command(self, cmd, instance, key, path, **kwargs):
+        cwd = '/tmp/'
+        env = {}
+        env['BORG_PASSPHRASE'] = key
+        env['BORG_REPO'] = path
+        env['BORG_RELOCATED_REPO_ACCESS_IS_OK'] = 'yes'
+        env['BORG_RSH'] = 'ssh -o StrictHostKeyChecking=no -o UserKnownHostsFile=/dev/null'
+        args = cmd
+        safe_args = args
+
+        kwargs['private_data_dir'] = self.build_private_data_dir(instance, **kwargs)
+        kwargs['private_data_files'] = self.build_private_data_files(instance, **kwargs)
+        kwargs['passwords'] = self.build_passwords(instance, **kwargs)
+
+        expect_passwords = {}
+        for k, v in self.get_password_prompts(**kwargs).items():
+            expect_passwords[k] = kwargs['passwords'].get(v, '') or ''
+
+        _kw = dict(
+            expect_passwords=expect_passwords,
+            job_timeout=getattr(settings, 'DEFAULT_JOB_TIMEOUT', 0),
+            idle_timeout=getattr(settings, 'JOB_RUN_IDLE_TIMEOUT', None),
+            extra_update_fields={},
+            pexpect_timeout=getattr(settings, 'PEXPECT_TIMEOUT', 5),
+        )
+        stdout_handle = StringIO()
+
+        ssh_key_path = self.get_ssh_key_path(instance, **kwargs)
+        # If we're executing on an isolated host, don't bother adding the
+        # key to the agent in this environment
+        if ssh_key_path:
+            ssh_auth_sock = os.path.join(kwargs['private_data_dir'], 'ssh_auth.sock')
+            args = run.wrap_args_with_ssh_agent(args, ssh_key_path, ssh_auth_sock)
+            safe_args = run.wrap_args_with_ssh_agent(safe_args, ssh_key_path, ssh_auth_sock)
+
+        status, rc = run.run_pexpect(
+            args, cwd, env, stdout_handle, **_kw
+        )
+
+        lines = stdout_handle.getvalue().splitlines()
+        return lines
 
     def cleanup_jobs(self):
         # Sanity check: Is there already a running job on the System?
@@ -40,21 +200,24 @@ class Command(BaseCommand):
                     for type in ('rootfs', 'vm', 'mysql', 'postgresql', 'config', 'piped', 'mail', 'folders'):
                         if '{}-'.format(type) in archive_name:
                             repoArchives.append(archive_name)
+                print(repoArchives)
 
                 entries = Job.objects.filter(job_type='job')
                 if entries.exists():
                     for entry in entries:
-                        if entry.archive_name != '' and entry.archive_name not in repoArchives:
+                        if entry.archive_name != '' and entry.archive_name and entry.archive_name not in repoArchives:
                             action_text = 'would delete' if self.dry_run else 'deleting'
-                            self.logger.info('%s %s', action_text, entry.archive_name)
+                            print('{} {}'.format(action_text, entry.archive_name))
                             if not self.dry_run:
+                                es.delete_by_query(index="catalog", doc_type='_doc',
+                                                   body={"query": {"match": {"archive_name": entry.archive_name}}})
                                 entry.delete()
+
         return 0, 0
 
     @transaction.atomic
     def handle(self, *args, **options):
         self.verbosity = int(options.get('verbosity', 1))
-        self.init_logging()
         self.dry_run = bool(options.get('dry_run', False))
 
         model_names = ('jobs',)
@@ -69,7 +232,7 @@ class Command(BaseCommand):
             if m in models_to_cleanup:
                 skipped, deleted = getattr(self, 'cleanup_%s' % m)()
                 if self.dry_run:
-                    self.logger.log(99, '%s: %d would be deleted, %d would be skipped.', m.replace('_', ' '),
-                                    deleted, skipped)
+                    print('{}: {} would be deleted, {} would be skipped.'.format(m.replace('_', ' '),
+                                    deleted, skipped))
                 else:
-                    self.logger.log(99, '%s: %d deleted, %d skipped.', m.replace('_', ' '), deleted, skipped)
+                    print('{}: {} deleted, {} skipped.'.format(m.replace('_', ' '), deleted, skipped))
